+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Fixed
 
+* `Client` now properly encode parameters sent in query strings.
+
 ### Deprecated
 
 ### Removed